SPLIT SCRIPT.JS IN TO SMALLER MODULES TO BE EASIER TO MANAGE
