+++ playerStart.js
@@ -1,0 +1,66 @@
+//Importing the fuel tank so that it can be attached to the game area.
+import { fuelTank } from './fuel.js';
+// Import question display. 
+import { questioning, questionsBox } from './qPicker.js';
+
+// Retrieving the game area so that playerStart can use it.
+const gameArea = document.querySelector('.gameArea');
+
+// Declaring global variables here so they can be exported and 
+//used once playerStart has set values to them.
+let playerName;
+let shipName;
+
+// Getting player and ship name before starting game.
+function playerStart() {
+    const launch = document.createElement('div');
+    launch.classList.add('launch');
+  
+    const nameLabel = document.createElement('label');
+    nameLabel.textContent = 'Captain! What should we call you?';
+    nameLabel.setAttribute('for', 'getName');
+    launch.appendChild(nameLabel);
+    const getName = document.createElement('input');
+    getName.setAttribute('name', 'getName');
+    launch.appendChild(getName);
+  
+    const shipLabel = document.createElement('label');
+    shipLabel.textContent = 'What shall we call our ship Captain?'
+    shipLabel.setAttribute('for', 'shipName');
+    launch.appendChild(shipLabel);
+    const getShip = document.createElement('input');
+    getShip.setAttribute('name', 'shipName');
+    launch.appendChild(getShip);
+  
+    const launchBtn = document.createElement('button');
+    launchBtn.textContent = 'Launch';
+    launch.appendChild(launchBtn);
+  
+    const rocket = new Image(50, 50);
+    rocket.src = 'img/rocket.svg';
+    launch.appendChild(rocket);
+  
+    const earth = new Image(100, 100);
+    earth.src = 'img/earth.svg';
+    launch.appendChild(earth);
+  
+    launchBtn.addEventListener('click', () => {
+      playerName = getName.value;
+      shipName = getShip.value;
+      launch.removeChild(nameLabel);
+      launch.removeChild(getName);
+      launch.removeChild(shipLabel);
+      launch.removeChild(getShip);
+      launch.removeChild(launchBtn);
+      rocket.classList.add('takeOff');
+      setTimeout(() => {
+        gameArea.removeChild(launch);
+        gameArea.appendChild(fuelTank);
+        questioning.displayQ();
+      }, 3000);
+    });
+  
+    gameArea.appendChild(launch);
+}
+
+export { playerStart, playerName, shipName }
